requirements.txt
flask
rembg
pillow
python-dotenv
onnxruntime  